style: consolidate ImageCropper tests into one PascalCase file, fix rest comment
- Merge the native-attr forwarding + onError tests into ImageCropper.test.tsx and
  drop image-cropper.test.tsx, matching the repo's ComponentName.test.tsx
  convention (the kebab file was the only exception).
- Correct the {...rest} comment: className is destructured out of rest, so spread
  order is irrelevant — the old comment implied an order-dependent guard.
